requirements.txt
streamlit
sqlalchemy
passlib
pyjwt
reportlab
pandas
matplotlib